[ICU-8474] Updated the plural rules factory to match CLDR; took a first pass at modifying NumberFormat to get the right information out for plural select with fractions. X-Commit-URL: https://ssl.icu-project.org/trac/changeset/33649 Former-commit-id: 6e1bf08aff7ce6052802c06375ae3652b0f3c95c

+++ icu4j/main/classes/core/src/com/ibm/icu/text/PluralRules.java
@@ -384,7 +384,7 @@ public class PluralRules implements Serializable {
         public final boolean hasIntegerValue;
         public final boolean isNegative;
 
-        public NumberInfo(double n, int v, int f) {
+        public NumberInfo(double n, int v, long f) {
             isNegative = n < 0;
             source = isNegative ? -n : n;
             visibleFractionDigitCount = v;
@@ -1357,10 +1357,11 @@ public class PluralRules implements Serializable {
      * @internal
      * @deprecated This API is ICU internal only.
      */
-    public String select(double number, int countVisibleFractionDigits, int fractionaldigits) {
+    public String select(double number, int countVisibleFractionDigits, long fractionaldigits) {
         return rules.select(new NumberInfo(number, countVisibleFractionDigits, fractionaldigits));
     }
 
+    
     /**
      * Given a number, returns the keyword of the first rule that applies to
      * the number.
